Split types into module and submodules
Without breaking the build because we reexport the private submodule
types in the crate::v1::types module.
