+++ librepology/src/v1/types/download.rs
@@ -1,0 +1,16 @@
+use std::ops::Deref;
+
+use url::Url;
+
+// list of package downloads
+#[derive(Eq, PartialEq, Clone, Debug, Serialize, Deserialize)]
+pub struct Download(#[serde(with = "url_serde")] Url);
+
+impl Deref for Download {
+    type Target = Url;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
