+++ librepology/src/v1/types/effname.rs
@@ -1,0 +1,14 @@
+use std::ops::Deref;
+
+// list of package downloads
+#[derive(Eq, PartialEq, Ord, PartialOrd, Clone, Debug, Serialize, Deserialize)]
+pub struct EffName(String);
+
+impl Deref for EffName {
+    type Target = String;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
