+++ librepology/src/v1/types/license.rs
@@ -1,0 +1,14 @@
+use std::ops::Deref;
+
+// list of package licenses
+#[derive(Eq, PartialEq, Ord, PartialOrd, Clone, Debug, Serialize, Deserialize)]
+pub struct License(String);
+
+impl Deref for License {
+    type Target = String;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
