+++ librepology/src/v1/types/name.rs
@@ -1,0 +1,14 @@
+use std::ops::Deref;
+
+// package name as in repository (if different from version)
+#[derive(Eq, PartialEq, Ord, PartialOrd, Clone, Debug, Serialize, Deserialize)]
+pub struct Name(String);
+
+impl Deref for Name {
+    type Target = String;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
